Fix #1799: Move Iterable to java.lang.
