+++ javalanglib/src/main/scala/java/lang/Iterable.scala
@@ -1,4 +1,6 @@
-package java.util
+package java.lang
+
+import java.util.Iterator
 
 trait Iterable[T] {
   def iterator(): Iterator[T]
